Fixed Bugs in schemas

+++ schemas/account.py
@@ -14,5 +14,5 @@ class Transfer(BaseModel):
 
     from_account_id: int
     to_account_id: int
-    amount: int
+    amount: float
